+++ src/main/java/com/daruda/darudaserver/domain/user/controller/AuthController.java
@@ -18,10 +18,10 @@ import com.daruda.darudaserver.domain.user.dto.response.LoginResponse;
 import com.daruda.darudaserver.domain.user.dto.response.SignUpSuccessResponse;
 import com.daruda.darudaserver.domain.user.dto.response.UserInfo;
 import com.daruda.darudaserver.domain.user.entity.enums.SocialType;
-import com.daruda.darudaserver.domain.user.service.LoginService;
+import com.daruda.darudaserver.domain.user.service.AuthService;
 import com.daruda.darudaserver.domain.user.service.SocialService;
-import com.daruda.darudaserver.domain.user.service.UserService;
 import com.daruda.darudaserver.global.annotation.DisableSwaggerSecurity;
+import com.daruda.darudaserver.global.auth.jwt.service.TokenService;
 import com.daruda.darudaserver.global.common.response.ApiResponse;
 import com.daruda.darudaserver.global.error.code.SuccessCode;
 
@@ -38,16 +38,17 @@ import lombok.extern.slf4j.Slf4j;
 @Tag(name = "auth 컨트롤러", description = "로그인과 관련된 API를 처리합니다.")
 @RequiredArgsConstructor
 public class AuthController {
-	private final UserService userService;
-	private final LoginService loginService;
+
+	private final AuthService authService;
+	private final TokenService tokenService;
 
 	@DisableSwaggerSecurity
 	@GetMapping("/login-url")
 	@Operation(summary = "소셜 로그인 URL 반환", description = "소셜 로그인 URL을 반환합니다.")
-	public ResponseEntity<ApiResponse<String>> requestLogin(
+	public ResponseEntity<ApiResponse<String>> requestLoginUrl(
 		@Parameter(description = "소셜 로그인 타입", example = "KAKAO")
 		@RequestParam SocialType socialType) {
-		SocialService socialService = loginService.findSocialService(socialType);
+		SocialService socialService = authService.findSocialService(socialType);
 		String redirectUrl = socialService.getLoginUrl();
 		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(redirectUrl, SuccessCode.SUCCESS_REDIRECT));
 	}
@@ -56,45 +57,45 @@ public class AuthController {
 	@PostMapping(value = "/login")
 	@Operation(summary = "소셜 로그인",
 		description = "소셜 로그인에서 발급받은 Authorization Code를 통해, 로그인을 진행합니다.")
-	public ResponseEntity<ApiResponse<LoginResponse>> postAuthenticationCode(
+	public ResponseEntity<ApiResponse<LoginResponse>> login(
 		@Parameter(description = "Authorization Code", example = "1234")
 		@RequestParam("code") String code,
 		@RequestBody LoginRequest loginRequest) {
 		log.debug("CODE = {}", code);
-		SocialService socialService = loginService.findSocialService(loginRequest.socialType());
+		SocialService socialService = authService.findSocialService(loginRequest.socialType());
 		UserInfo userInfo = socialService.getInfo(code);
-		LoginResponse loginResponse = userService.oauthlogin(userInfo);
-		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(loginResponse, SuccessCode.SUCCESS_CREATE));
+		LoginResponse loginResponse = authService.login(userInfo);
+		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(loginResponse, SuccessCode.SUCCESS_LOGIN));
 	}
 
 	@DisableSwaggerSecurity
 	@PostMapping("/sign-up")
 	@Operation(summary = "회원 가입", description = "회원 가입을 진행합니다.")
-	public ResponseEntity<ApiResponse<SignUpSuccessResponse>> signUp(@Valid @RequestBody SignUpRequest signUpRequest) {
-		SignUpSuccessResponse signUpSuccessResponse = userService.createUser(signUpRequest.email(),
+	public ResponseEntity<ApiResponse<SignUpSuccessResponse>> register(
+		@Valid @RequestBody SignUpRequest signUpRequest) {
+		SignUpSuccessResponse signUpSuccessResponse = authService.register(signUpRequest.email(),
 			signUpRequest.nickname(), signUpRequest.positions());
-		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(signUpSuccessResponse, SuccessCode.SUCCESS_LOGIN));
+		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(signUpSuccessResponse, SuccessCode.SUCCESS_CREATE));
 	}
 
 	@PostMapping("/logout")
 	@Operation(summary = "로그아웃", description = "로그아웃을 진행합니다.")
-	public ResponseEntity<ApiResponse<Long>> logOut(@AuthenticationPrincipal Long userId) {
-		Long returnedUserId = userService.deleteUser(userId);
-		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(returnedUserId, SuccessCode.SUCCESS_LOGOUT));
+	public ResponseEntity<ApiResponse<Long>> logout(@AuthenticationPrincipal Long userId) {
+		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(authService.logout(userId), SuccessCode.SUCCESS_LOGOUT));
 	}
 
 	@DisableSwaggerSecurity
 	@PostMapping("/reissue")
 	@Operation(summary = "Access Token 재발급", description = "Refresh Token을 통해 Access Token을 재발급합니다.")
-	public ResponseEntity<ApiResponse<JwtTokenResponse>> regenerateToken(@RequestBody ReissueTokenRequest request) {
-		JwtTokenResponse jwtTokenResponse = userService.reissueToken(request.refreshToken());
-		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(jwtTokenResponse, SuccessCode.SUCCESS_REISSUE));
+	public ResponseEntity<ApiResponse<JwtTokenResponse>> reissueToken(@RequestBody ReissueTokenRequest request) {
+		return ResponseEntity.ok(ApiResponse.ofSuccessWithData(tokenService.reissueToken(request.refreshToken()),
+			SuccessCode.SUCCESS_REISSUE));
 	}
 
 	@DeleteMapping("/withdraw")
 	@Operation(summary = "회원 탈퇴", description = "회원 탈퇴를 진행합니다.")
-	public ResponseEntity<ApiResponse<?>> withdrawUser(@AuthenticationPrincipal Long userId) {
-		userService.withdrawMe(userId);
+	public ResponseEntity<ApiResponse<?>> withdraw(@AuthenticationPrincipal Long userId) {
+		authService.withdraw(userId);
 		return ResponseEntity.ok(ApiResponse.ofSuccess(SuccessCode.SUCCESS_WITHDRAW));
 	}
 }
